Improve pre-commit sample for openwhisk
The current code sample does not scan the full openwhisk directory but only starting from the current directory. In addition, no configuration file is specified.

With this change, the pre-commit hook scans the full openwhisk directory and uses the scan configuration that is also used in the openwhisk Travis configuration.